requirements.txt
pymongo[srv]
passlib
bcrypt==3.2.0
customtkinter
opencv-python
face_recognition
Pillow
numpy
mtcnn
tensorflow